Update pka_marker.py
The results file now defaults (with appropriate settings in .env.shared) to using the name of the PKA directory and instances of Guest PKA submissions are recorded in a separate guest.txt file.  This allows the CSV file to be uploaded to the VLE without having to remove any instances of Guest submissions.

Command line options for --no-guest-file to disable the use of the guest file and put the resulting Guest scores of 0 into the results CSV and --use-default-results-file added to override a False value for USE_DEFAULT_RESULTS_FILE (.env.shared).

Mutually excluse command line options are now grouped as such using argparse.  A conditional still covers the case where argparse does not support grouping of different types such as a string and a boolean for --results-file and --no-csv.